requirements.txt
llama-index==0.10.15
openai==1.13.3
python-dotenv==1.0.1
